requirements.txt
pyautogui==0.9.54
opencv-python==4.8.0.76
numpy==1.24.3
keyboard==0.13.5
Pillow==10.0.0 